Install: fix postinstall script on osx
The macOS install encounteres the following error when running the postinstall script:

    ./postinstall: /tmp/PKInstallSandbox.qGkbqk/Scripts/com.microsoft.GitCredentialManager.kjGWk5/postinstall: line 21: [: missing `]'

This resulted in the credential helper not being correctly configured.

The problem is with the test on the indicated line - the '[ ]' test does not
support '&&' statements inside the test expression, as it splits the expression into separate
commands.