ui: correctly check for GUI on Win/Mac/POSIX
Replace the `PlatformUtils.IsDesktopSession` util method with platform
specific components that check use native APIs to determine the state of
the current session.

`Environment.UserInteractive` is hard-coded to return true for POSIX and
Windows platforms on .NET Core 2.x and 3.x.
https://github.com/dotnet/runtime/issues/770

The .NET 5 implementation (not yet released) fixes this for the Windows
platform only. We take a copy of that implementation for Windows.

On macOS we use the `SessionGetInfo` from the Security.framework.

On POSIX in general we also check for the X11 $DISPLAY environment
variable.