build: workaround Azure Pipelines build bug
For some reason our Azure Pipelines builds are failing (occasionally) to
restore packages _unless_ we explicitly run a `dotnet restore` before
our `dotnet build`. The Azure Pipelines team are still not sure why this
is failing, or why it only happens one some of the hosted agents and not
others (despite compiling the same commit)!