hostprovider: update host prvdr base to use service name
Update the HostProvider base class to use the new 'service name'
abstraction rather than the simple 'unique credential key' one.

With this model we can better issue credential storage queries where the
username may not be specified explicitly in a get request (often the
case as the username is not always included in the remote URL for many
services).