oauth: disconnect streams from xdg-open
When using the `Process` class to open the user's default browser on
Linux, utilities like `xdg-open` are used. Some of these utilities
do not disconnect child processes from our standard input/output/error
streams. At the same time, browsers like Chromium like to write to
stdout and stderr, which gets fed back to Git or the user's terminal,
respectively. The latter looks messy, and the former causes Git to fail.

On Linux, we instead manually locate a suitable 'shell execute' utility
and launch them directly - this way we can redirect the standard
output/error streams.

For Windows and macOS, this is not an issue and we continue to use the
Framework code to do 'shell execute'.