osx: update the Mac installer to make product archives
Update the macOS installer build to also create a distribution/product
archive package, that contains the flat component package we were
already building.

The product archive allows us to do things like set welcome and
conclusion messages, titles, and display a license (like the Windows
installers do).

In the new welcome and conclusion screens we include helpful links and
instructions on how to uninstall GCM Core, and how to configure it for
other users on the system.

The macOS installer continues to install in a global location, but only
configure the user's global Git configuration.