config: fix reading of empty string Git config values
Previously trying to read a single Git configuration setting that had
the value of "" empty string resulted in returning `false` in our
`TryGet` method.

Move to using the null-terminating output from Git for easier parsing,
just like we use for the other read-commands.

The `GetAll` method didn't have this issue as it was already using
`--null`.