azrepos: add non-PAT based auth method to AzRepos
Add a new settings to the Azure Repos provider that instructs GCM to
return the Azure access token directly, rather than use that token to
generate a new Azure DevOps Personal Access Token (PAT).

At the moment the only indication as to what user account a user wants
to use is via the userinfo parts of the remote URL. This must be set
manually.

The default configuration is to continue to use PATs, for now.