msauth: include better warning message on Keychain errors
Sometimes the login Keychain is not accessible and the MSAL.Extensions
library cannot read/write the token cache. The specific error
encountered internally is `errSecAuthFailed` - neither we nor the MSAL
team know why this happens sometimes.

Locking and then unlocking the login Keychain manually, or simply
restarting the computer appears to fix the problem.

Include the lock+unlock/restart remedy in the warning message on macOS.