PosixTerminal: split macOS and Linux impl and fix
Previously the PosixTerminal class used the POSIX termios structure and
tcgetattr/tcsetattr functions for both macOS and Linux systems.

However, it transpires that the definition of the termios flags fields
differs between Mac and Linux. On Linux the fields are unsigned int (4
bytes), but on macOS the fields are unsigned long (8 bytes).

This change splits the termios structure into Mac and Linux flavours and
sets the correct field offsets.