github: show terminal prompts when no GUI
Show the terminal/TTY based prompts when there is no desktop/GUI
sessions present for GitHub.

Currently the Bitbucket and Microsoft auth stacks correctly detect the
lack of a desktop session and use terminal prompts - GitHub does not.

Without this change, SSH-ing in to a machine we'd still try and show a
UI prompt where there is no way to interact with it!